change progress to next level settings and update the activity level when reached level 10

+++ lib/achievement/screen/physio/patient_details_screen.dart
@@ -104,7 +104,7 @@ class _PatientDetailsScreenState extends State<PatientDetailsScreen> {
     if (userSnapshot.exists) {
       // Retrieve the existing experience and level from the document snapshot
       level = userSnapshot.get('level') ?? 0;
-      progressToNextLevel = userSnapshot.get('progressToNextLevel') ?? 0.0;
+      progressToNextLevel = userSnapshot.get('progressToNextLevel').toDouble() ?? 0.0;
     }
     final CollectionReference ptCollection =
         userCollection.doc(uid).collection('pt_activities');
