+++ lib/profile/screen/profile_screen.dart
@@ -342,7 +342,7 @@ class _ProfileScreenState extends State<ProfileScreen> {
                                             lineHeight: 10.0,
                                             animationDuration: 2000,
                                             percent:
-                                                userData['progressToNextLevel'],
+                                                userData['progressToNextLevel'].toDouble(),
                                             barRadius: Radius.circular(10.0),
                                             progressColor: Colors.yellow,
                                             padding: EdgeInsets.zero,
